+++ tests/test_py311_generated/test_new_style_annotations_min_py312_generated.py
@@ -458,14 +458,14 @@ def test_generic_type_alias_individual_resolution() -> None:
     This verifies that the type parameter resolution works correctly for
     individual generic type aliases before testing unions.
     """
-    from tyro._resolver import TypeParamResolver
+    from tyro._resolver import TypeParamAndAliasResolver
 
     # Test individual resolution
     alias_int = ConstrainedTuple[int]
     alias_bool = ConstrainedTuple[bool]
 
-    resolved_int = TypeParamResolver.concretize_type_params(alias_int)  # type: ignore
-    resolved_bool = TypeParamResolver.concretize_type_params(alias_bool)  # type: ignore
+    resolved_int = TypeParamAndAliasResolver.resolve_recursive(alias_int)  # type: ignore
+    resolved_bool = TypeParamAndAliasResolver.resolve_recursive(alias_bool)  # type: ignore
 
     # Check that the resolved types are what we expect
 
@@ -479,11 +479,11 @@ def test_generic_type_alias_union_resolution() -> None:
     This tests the internal type resolution mechanism that was buggy.
     """
 
-    from tyro._resolver import TypeParamResolver
+    from tyro._resolver import TypeParamAndAliasResolver
 
     # Test union resolution
     union_type = ConstrainedTuple[int] | ConstrainedTuple[bool]
-    resolved_union = TypeParamResolver.concretize_type_params(union_type)  # type: ignore
+    resolved_union = TypeParamAndAliasResolver.resolve_recursive(union_type)  # type: ignore
 
     # The resolved union should contain all four tuple types
     expected = tuple[int] | tuple[int, int] | tuple[bool] | tuple[bool, bool]
